Refactor EditUserForm
